Refactored Skill And Personnel Management Into Utility Class
- Moved skill, loyalty, and advantage-related methods from `HireBulkPersonnelDialog` to a new `PersonUtility` class for better modularity and reusability.
- Replaced integer-based experience levels with `SkillLevel` enums across affected classes for improved readability and type safety.
- Updated references to the refactored methods in impacted classes (`HireBulkPersonnelDialog`, `UnitTableMouseAdapter`, `AtBContract`).
- Deprecated `SkillLevel.getAdjustedValue()` in favor of the newly added `getExperienceLevel()` method.
- Added Javadoc documentation to `PersonUtility` methods for clear descriptions of functionality.
